// generated/requests/types.gen.ts
// This file is auto-generated by @hey-api/openapi-ts

export type ClientOptions = {
    baseUrl: 'https://dummyjson.com' | (string & {});
};

export type Product = {
    id?: number;
    title?: string;
    price?: number;
    description?: string;
    category?: string;
    thumbnail?: string;
    rating?: number;
    stock?: number;
};

export type GetProductsData = {
    body?: never;
    path?: never;
    query?: {
        limit?: number;
    };
    url: '/products';
};

export type GetProductsResponses = {
    /**
     * Products list
     */
    200: {
        products?: Array<Product>;
    };
};

export type GetProductsResponse = GetProductsResponses[keyof GetProductsResponses];

export type GetProductByIdData = {
    body?: never;
    path: {
        id: number;
    };
    query?: never;
    url: '/products/{id}';
};

export type GetProductByIdResponses = {
    /**
     * Single product
     */
    200: Product;
};

export type GetProductByIdResponse = GetProductByIdResponses[keyof GetProductByIdResponses];